Move universal converters to `bot/converters.py`.
